requirements.txt
Django==1.9
Markdown==2.6.5
# Pillow==3.0.0
PyDictionary==1.5.2
Werkzeug==0.11.2
argparse==1.2.1
beautifulsoup4==4.4.1
click==6.2
decorator==4.0.6
django-admin-bootstrapped==2.5.6
django-authtools==1.4.0
django-braces==1.8.1
django-crispy-forms==1.5.2
django-debug-toolbar==1.4
django-environ==0.4.0
django-filter==0.11.0
djangorestframework==3.3.1
easy-thumbnails==2.3
futures==3.0.3
goslate==1.5.0
ipdb==0.8.1
ipython==4.0.1
ipython-genutils==0.1.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
ptyprocess==0.5
requests==2.8.1
simplegeneric==0.8.1
six==1.10.0
smtpapi==0.2.0
sqlparse==0.1.18
traitlets==4.0.0
